flask-multi-db-monorepo/unified_app/requirements.txt
# Unified Multi-Database Demo Application
# Flask and core
flask>=3.0.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0

# Authentication
bcrypt>=4.0.0
flask-session>=0.5.0

# Database Drivers
mssql-python>=1.3.0   # Azure SQL with native Azure AD support
psycopg2-binary>=2.9.9   # PostgreSQL
pymongo>=4.6.0   # MongoDB / Azure Cosmos DB for MongoDB

# Azure OpenAI for embeddings and NL queries
openai>=1.12.0

# Azure Cosmos DB NoSQL for activity tracking
azure-cosmos>=4.7.0

# Utilities
requests>=2.31.0
